+++ lib/glimmer/swt/image_proxy.rb
@@ -83,6 +83,11 @@ module Glimmer
           @swt_image = Image.new(*@args)
           @original_image_data = @image_data = @swt_image.image_data
         end
+        @swt_image.singleton_class.alias_method(:dispose_without_glimmer, :dispose)
+        @swt_image.singleton_class.define_method(:dispose) {
+          clear_shapes
+          dispose_without_glimmer
+        }
         post_add_content if content.nil?
       end