+++ lib/glimmer/swt/widget_proxy.rb
@@ -165,6 +165,9 @@ module Glimmer
           DEFAULT_INITIALIZERS[underscored_widget_name.to_s.to_sym]&.call(@swt_widget)
           @parent_proxy.post_initialize_child(self)
         end
+        on_widget_disposed {
+          clear_shapes
+        }
       end
       
       # Subclasses may override to perform post initialization work on an added child